Fixed #2 in a way that I think will be intuitive
There are two bounding box methods.
The getBoundingBox never returns null and creates the box if it does
not exist.
the getRawBoundingBox returns the current object or a clone of it and
may return null.
This second method can be used to prevent possible infinite loops and
as a result
such is only available to objects that override the calculateBBBox
method